+++ Tether/Program.cs
@@ -1,7 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.IO;
 using System.Linq;
+using System.Management;
 using System.Reflection;
 using System.Threading.Tasks;
 using NLog;
@@ -9,6 +11,7 @@ using Quartz;
 using Tether.Plugins;
 using Topshelf;
 using Topshelf.Quartz;
+using Utilities.DataTypes.ExtensionMethods;
 
 namespace Tether
 {
@@ -22,6 +25,169 @@ namespace Tether
             CheckTypes = new Dictionary<string, ICheck>();
             slices = new List<string>();
         }
+
+        public dynamic GetSlice(string Name)
+        {
+            MethodInfo method = GetType().GetMethod("PopulateMultiple", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Static).MakeGenericMethod(new Type[] { Type.GetType(Name) });
+
+            var invoke = method.Invoke(this, null) as dynamic;
+
+            return invoke;
+        }
+
+        private static List<T> PopulateMultiple<T>() where T : new()
+        {
+            var t = new List<T>();
+            var pcga = typeof(T).Attribute<PerformanceCounterGroupingAttribute>();
+
+            if (pcga != null)
+            {
+
+                if (pcga.UsePerformanceCounter)
+                {
+                    var PerfCounter = new PerformanceCounterCategory(pcga.WMIClassName);
+
+                    var instances = PerfCounter.GetInstanceNames().PerformCounterFiltering(pcga.Selector, pcga.SelectorValue, pcga.ExclusionContains);
+
+                    foreach (var instance in instances)
+                    {
+                        var item = new T();
+
+                        IEnumerable<string> names = typeof(T).GetProperties()
+                                                .Where(f => f.Attribute<PerformanceCounterValueExcludeAttribute>() == null)
+                                                .Select(
+                                                    delegate (PropertyInfo info)
+                                                    {
+                                                        if (info.Attribute<PerformanceCounterValueAttribute>() != null && info.Attribute<PerformanceCounterValueAttribute>().PropertyName != null)
+                                                        {
+                                                            return info.Attribute<PerformanceCounterValueAttribute>().PropertyName;
+                                                        }
+                                                        return info.Name;
+                                                    });
+
+                        foreach (var name in names)
+                        {
+                            PropertyInfo property = typeof(T).GetProperties()
+                                    .FirstOrDefault(
+                                        f => (f.Attribute<PerformanceCounterValueAttribute>() != null && f.Attribute<PerformanceCounterValueAttribute>().PropertyName == name) ||
+                                        f.Name == name && f.Attribute<PerformanceCounterValueExcludeAttribute>() == null);
+
+
+                            if (property.Attribute<PerformanceCounterInstanceNameAttribute>() != null)
+                            {
+                                property.SetValue(item, instance, null);
+                            }
+                            else
+                            {
+                                var performanceCounters = PerfCounter.GetCounters(instance);
+                                try
+                                {
+
+                                    var changeType = Convert.ChangeType(performanceCounters.FirstOrDefault(e => e.CounterName == name).NextValue(), property.PropertyType);
+
+                                    if (property.Attribute<PerformanceCounterValueAttribute>() != null && property.Attribute<PerformanceCounterValueAttribute>().Divisor > 0)
+                                    {
+                                        if (property.PropertyType == typeof(long))
+                                        {
+                                            changeType = (long)changeType / property.Attribute<PerformanceCounterValueAttribute>().Divisor;
+                                        }
+                                        else if (property.PropertyType == typeof(int))
+                                        {
+                                            changeType = (int)changeType / property.Attribute<PerformanceCounterValueAttribute>().Divisor;
+                                        }
+                                        else if (property.PropertyType == typeof(short))
+                                        {
+                                            changeType = (short)changeType / property.Attribute<PerformanceCounterValueAttribute>().Divisor;
+                                        }
+                                    }
+
+                                    property.SetValue(item, changeType, null);
+                                }
+                                catch (Exception e)
+                                {
+                                    logger.ErrorException("Error on property " + name, e);
+                                }
+                                finally
+                                {
+                                    DisposeAll(performanceCounters);
+
+                                }
+                            }
+
+                        }
+                        t.Add(item);
+
+                    }
+                    return t;
+
+                }
+
+
+                var searcher = new ManagementObjectSearcher(pcga.WMIRoot, "SELECT * FROM " + pcga.WMIClassName);
+
+                foreach (ManagementObject var in searcher.Get().Cast<ManagementObject>().PerformFiltering(pcga.Selector, pcga.SelectorValue, pcga.ExclusionContains, pcga.Subquery))
+                {
+                    var item = new T();
+
+                    IEnumerable<string> names = typeof(T).GetProperties()
+                        .Where(f => f.Attribute<PerformanceCounterValueExcludeAttribute>() == null)
+                        .Select(
+                            delegate (PropertyInfo info)
+                            {
+                                if (info.Attribute<PerformanceCounterValueAttribute>() != null && info.Attribute<PerformanceCounterValueAttribute>().PropertyName != null)
+                                {
+                                    return info.Attribute<PerformanceCounterValueAttribute>().PropertyName;
+                                }
+                                return info.Name;
+                            });
+
+                    foreach (var name in names)
+                    {
+                        var property = typeof(T).GetProperties()
+                                .FirstOrDefault(
+                                    f => (f.Attribute<PerformanceCounterValueAttribute>() != null && f.Attribute<PerformanceCounterValueAttribute>().PropertyName == name) || f.Name == name && f.Attribute<PerformanceCounterValueExcludeAttribute>() == null);
+
+                        try
+                        {
+
+                            var changeType = Convert.ChangeType(var[name], property.PropertyType);
+
+                            if (property.Attribute<PerformanceCounterValueAttribute>() != null && property.Attribute<PerformanceCounterValueAttribute>().Divisor > 0)
+                            {
+                                if (property.PropertyType == typeof(long))
+                                {
+                                    changeType = (long)changeType / property.Attribute<PerformanceCounterValueAttribute>().Divisor;
+                                }
+                                else if (property.PropertyType == typeof(int))
+                                {
+                                    changeType = (int)changeType / property.Attribute<PerformanceCounterValueAttribute>().Divisor;
+                                }
+                                else if (property.PropertyType == typeof(short))
+                                {
+                                    changeType = (short)changeType / property.Attribute<PerformanceCounterValueAttribute>().Divisor;
+                                }
+                            }
+
+
+
+                            property.SetValue(item, changeType, null);
+                        }
+                        catch (Exception e)
+                        {
+                            logger.Error(e, $"Error on property {name}");
+                        }
+
+                    }
+                    t.Add(item);
+                }
+
+
+            }
+
+
+            return t;
+        }
+
 
         public dynamic PerformCheck(string checkName)
         {
